- move generic method to BaseServiceImpl.
