+++ src/main/java/com/commercetools/project/sync/service/impl/BaseServiceImpl.java
@@ -1,8 +1,19 @@
 package com.commercetools.project.sync.service.impl;
 
+import static java.util.stream.Collectors.toSet;
+import static org.apache.commons.lang3.StringUtils.isBlank;
+
+import com.commercetools.project.sync.model.ResourceIdsGraphQlRequest;
+import com.commercetools.sync.commons.models.GraphQlQueryResources;
+import com.commercetools.sync.commons.models.ResourceKeyId;
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import io.sphere.sdk.client.SphereClient;
+import io.sphere.sdk.models.WithKey;
+import java.util.List;
+import java.util.Set;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionStage;
 import javax.annotation.Nonnull;
 
 public class BaseServiceImpl {
@@ -23,5 +34,43 @@ public class BaseServiceImpl {
 
   protected SphereClient getCtpClient() {
     return ctpClient;
+  }
+
+  protected <T extends WithKey> CompletionStage<List<T>> fetchAndFillReferenceIdToKeyCache(
+      @Nonnull final List<T> resourceList,
+      final Set<String> ids,
+      final GraphQlQueryResources requestType) {
+    final Set<String> nonCachedReferenceIds = getNonCachedReferenceIds(ids);
+    if (nonCachedReferenceIds.isEmpty()) {
+      return CompletableFuture.completedFuture(resourceList);
+    }
+
+    return getCtpClient()
+        .execute(new ResourceIdsGraphQlRequest(nonCachedReferenceIds, requestType))
+        .toCompletableFuture()
+        .thenApply(
+            results -> {
+              cacheProductTypeKeys(results.getResults());
+              return resourceList;
+            });
+  }
+
+  @Nonnull
+  private Set<String> getNonCachedReferenceIds(@Nonnull final Set<String> referenceIds) {
+    return referenceIds
+        .stream()
+        .filter(id -> null == referenceIdToKeyCache.getIfPresent(id))
+        .collect(toSet());
+  }
+
+  private void cacheProductTypeKeys(final Set<ResourceKeyId> results) {
+    results.forEach(
+        resourceKeyId -> {
+          final String key = resourceKeyId.getKey();
+          final String id = resourceKeyId.getId();
+          if (!isBlank(key)) {
+            referenceIdToKeyCache.put(id, key);
+          }
+        });
   }
 }
